requirements.txt
Django==1.7.5
django-multiurl==1.0.1
